[TASK] Refactor menu viewhelpers
Refacors all menu ViewHelpers to use much more of the TYPO3 core when retrieving page links and records, delegating most of the fetching to PageRepository.
