+++ Classes/ViewHelpers/Page/Menu/DeferredViewHelper.php
@@ -16,52 +16,11 @@ namespace FluidTYPO3\Vhs\ViewHelpers\Page\Menu;
  * TRUE - this will cause the output of the parent to only
  * contain the content of this ViewHelper.
  *
+ * DEPRECATED: use v:menu.deferred instead
+ *
  * @author Claus Due <claus@namelesscoder.net>
  * @package Vhs
  * @subpackage ViewHelpers\Page
  */
-class DeferredViewHelper extends AbstractMenuViewHelper {
-
-	/**
-	 * Initialize
-	 * @return void
-	 */
-	public function initializeArguments() {
-		parent::initializeArguments();
-		$this->overrideArgument('as', 'string', 'If used, stores the menu pages as an array in a variable named according to this value and renders the tag content - which means automatic rendering is disabled if this attribute is used', FALSE, NULL);
-	}
-
-	/**
-	 * @return string
-	 * @throws \Exception
-	 */
-	public function render() {
-		$as = $this->arguments['as'];
-		if (FALSE === $this->viewHelperVariableContainer->exists('FluidTYPO3\Vhs\ViewHelpers\Page\Menu\AbstractMenuViewHelper', 'deferredArray')) {
-			return NULL;
-		}
-		if (FALSE === $this->viewHelperVariableContainer->exists('FluidTYPO3\Vhs\ViewHelpers\Page\Menu\AbstractMenuViewHelper', 'deferredString')) {
-			return NULL;
-		}
-		if (NULL === $as) {
-			$content = $this->viewHelperVariableContainer->get('FluidTYPO3\\Vhs\\ViewHelpers\\Page\\Menu\\AbstractMenuViewHelper', 'deferredString');
-			$this->unsetDeferredVariableStorage();
-			return $content;
-		} elseif (TRUE === empty($as)) {
-			throw new \Exception('An "as" attribute was used but was empty - use a proper string value', 1370096373);
-		}
-		if (TRUE === $this->templateVariableContainer->exists($as)) {
-			$backupVariable = $this->templateVariableContainer->get($as);
-			$this->templateVariableContainer->remove($as);
-		}
-		$this->templateVariableContainer->add($as, $this->viewHelperVariableContainer->get('FluidTYPO3\\Vhs\\ViewHelpers\\Page\\Menu\\AbstractMenuViewHelper', 'deferredArray'));
-		$this->unsetDeferredVariableStorage();
-		$content = $this->renderChildren();
-		$this->templateVariableContainer->remove($as);
-		if (TRUE === isset($backupVariable)) {
-			$this->templateVariableContainer->add($as, $backupVariable);
-		}
-		return $content;
-	}
-
+class DeferredViewHelper extends \FluidTYPO3\Vhs\ViewHelpers\Menu\DeferredViewHelper {
 }
