+++ Tests/Unit/ViewHelpers/Page/LinkViewHelperTest.php
@@ -9,6 +9,7 @@ namespace FluidTYPO3\Vhs\ViewHelpers\Page;
  */
 
 use FluidTYPO3\Vhs\Tests\Unit\ViewHelpers\AbstractViewHelperTest;
+use TYPO3\CMS\Frontend\Page\PageRepository;
 
 /**
  * @protection on
@@ -17,24 +18,25 @@ use FluidTYPO3\Vhs\Tests\Unit\ViewHelpers\AbstractViewHelperTest;
 class LinkViewHelperTest extends AbstractViewHelperTest {
 
 	/**
+	 * @var PageRepository
+	 */
+	protected $pageRepository;
+
+	/**
 	 * @return void
 	 */
 	public function setUp() {
 		parent::setUp();
-		$GLOBALS['TYPO3_DB'] = $this->getMock(
-			'TYPO3\\CMS\\Core\\Database\\DatabaseConnection',
-			array('exec_SELECTquery', 'sql_fetch_assoc'),
-			array(), '', FALSE
-		);
-		$GLOBALS['TYPO3_DB']->expects($this->any())->method('exec_SELECTquery')->willReturn(NULL);
+		$this->pageRepository = $this->getMock(PageRepository::class, array('getPage'));
+		$GLOBALS['TSFE'] = (object) array('sys_page' => $this->pageRepository);
 	}
 
 	/**
 	 * @test
 	 */
 	public function generatesPageLinks() {
+		$this->pageRepository->expects($this->once())->method('getPage')->willReturn(array('uid' => '1', 'title' => 'test'));
 		$arguments = array('pageUid' => 1);
-		$GLOBALS['TYPO3_DB']->expects($this->any())->method('sql_fetch_assoc')->willReturn(array('uid' => '1', 'title' => 'test'));
 		$result = $this->executeViewHelper($arguments, array(), NULL, 'Vhs');
 		$this->assertNotEmpty($result);
 	}
@@ -44,7 +46,7 @@ class LinkViewHelperTest extends AbstractViewHelperTest {
 	 */
 	public function generatesNullLinkOnZeroPageUid() {
 		$arguments = array('pageUid' => 0);
-		$GLOBALS['TYPO3_DB']->expects($this->any())->method('sql_fetch_assoc')->willReturn(FALSE);
+		$this->pageRepository->expects($this->once())->method('getPage')->willReturn(NULL);
 		$result = $this->executeViewHelper($arguments, array(), NULL, 'Vhs');
 		$this->assertNull($result);
 	}
@@ -53,8 +55,8 @@ class LinkViewHelperTest extends AbstractViewHelperTest {
 	 * @test
 	 */
 	public function generatesPageLinksWithCustomTitle() {
+		$this->pageRepository->expects($this->never())->method('getPage');
 		$arguments = array('pageUid' => 1, 'pageTitleAs' => 'title');
-		$GLOBALS['TYPO3_DB']->expects($this->any())->method('sql_fetch_assoc')->willReturn(array('uid' => '1', 'title' => 'test'));
 		$result = $this->executeViewHelperUsingTagContent('Text', 'customtitle', $arguments, array(), 'Vhs');
 		$this->assertContains('customtitle', $result);
 	}
@@ -63,8 +65,8 @@ class LinkViewHelperTest extends AbstractViewHelperTest {
 	 * @test
 	 */
 	public function generatesPageWizardLinks() {
+		$this->pageRepository->expects($this->never())->method('getPage');
 		$arguments = array('pageUid' => '1 2 3 4 5 foo=bar&baz=123');
-		$GLOBALS['TYPO3_DB']->expects($this->any())->method('sql_fetch_assoc')->willReturn(array('uid' => '1', 'title' => 'test'));
 		$result = $this->executeViewHelper($arguments, array(), NULL, 'Vhs');
 		$this->assertNotEmpty($result);
 	}
