+++ Tests/Unit/ViewHelpers/Page/RootlineViewHelperTest.php
@@ -9,6 +9,7 @@ namespace FluidTYPO3\Vhs\ViewHelpers\Page;
  */
 
 use FluidTYPO3\Vhs\Tests\Unit\ViewHelpers\AbstractViewHelperTest;
+use TYPO3\CMS\Frontend\Page\PageRepository;
 
 /**
  * @protection off
@@ -18,6 +19,8 @@ use FluidTYPO3\Vhs\Tests\Unit\ViewHelpers\AbstractViewHelperTest;
 class RootlineViewHelperTest extends AbstractViewHelperTest {
 
 	public function testRender() {
+		$pageRepository = $this->getMock(PageRepository::class, array('dummy'));
+		$GLOBALS['TSFE'] = (object) array('sys_page' => $pageRepository);
 		$GLOBALS['TYPO3_DB'] = $this->getMock('TYPO3\\CMS\\Core\\Database\\DatabaseConnection', array('exec_SELECTgetSingleRow'), array(), '', FALSE);
 		$GLOBALS['TYPO3_DB']->expects($this->any())->method('exec_SELECTgetSingleRow')->willReturn(FALSE);
 		$this->assertEmpty($this->executeViewHelper());
